+++ lib/widgets/node_widget.dart
@@ -278,6 +278,7 @@ class NodeWidget<T> extends StatelessWidget {
       interactive: tooltipTheme.interactive ?? false,
       waitDuration: tooltipTheme.waitDuration,
       showDuration: tooltipTheme.showDuration,
+      hideOnEmptyMessage: tooltipTheme.hideOnEmptyMessage ?? true,
       message: resolvedBuilder == null ? message : null,
       tooltipBuilder: resolvedBuilder,
       child: child,